PR4 - Change link library to local copy
