+++ ui/app/components/account-dropdowns.js
@@ -2,7 +2,7 @@ const Component = require('react').Component
 const PropTypes = require('react').PropTypes
 const h = require('react-hyperscript')
 const actions = require('../actions')
-const genAccountLink = require('etherscan-link').createAccountLink
+const genAccountLink = require('../../lib/ubiqscan-link').createAccountLink
 const connect = require('react-redux').connect
 const Dropdown = require('./dropdown').Dropdown
 const DropdownMenuItem = require('./dropdown').DropdownMenuItem
