+++ ui/lib/ubiqscan-link/account-link.js
@@ -1,0 +1,8 @@
+const prefixForNetwork = require('./prefix-for-network')
+
+module.exports = function (address, network) {
+  const net = parseInt(network)
+  const prefix = prefixForNetwork(network)
+  return `http://ubiqscan.io/address/${address}?node=${prefix}`
+}
+
