+++ ui/lib/ubiqscan-link/explorer-link.js
@@ -1,0 +1,6 @@
+const prefixForNetwork = require('./prefix-for-network')
+
+module.exports = function (hash, network) {
+  const prefix = prefixForNetwork(network)
+  return `http://ubiqscan.io/tx/${hash}?node=${prefix}`
+}
