+++ ui/lib/ubiqscan-link/index.js
@@ -1,0 +1,10 @@
+const createExplorerLink = require('./explorer-link')
+const createAccountLink = require('./account-link')
+const createTokenLink = require('./token-link')
+
+module.exports = {
+  createExplorerLink,
+  createAccountLink,
+  createTokenLink
+}
+
